# speed up _assemble_generated_quantities