#### Submission Checklist

- [ ] Run unit tests
- [x] Declare copyright holder and open-source license: see below

#### Summary
```python
import numpy as np
import pandas as pd

def new(f):
    return pd.read_csv(
        f,
        sep=',',
        comment='#',
        header=0,
    ).values

# current implementation
def current(f):
    with open(f, 'r') as fd:
        lines = (line for line in fd if not line.startswith('#'))
        return  np.loadtxt(
            lines, dtype=np.ndarray, ndmin=2, skiprows=1, delimiter=','
        )
```

Runtime comparison for different file sizes on my laptop measured by ipython's `%timeit`:

|        | 49KB    | 2.2MB   | 66MB    | 132MB   |
|--------|---------|---------|---------|---------|
|new | 1.56 ms | 26.7 ms | 632 ms  | 1250 ms |
|current | 5.03 ms | 204 ms  | 2030 ms | 4190 ms |

The 49KB file is the one generated by the bernoulli experiment in the cmdstan repository.

Adding the parameters `quoting=3` or `low_memory=False` to `read_csv` did not improve the performance of `new`.

One difference between the two versions is, that for a file with zero samples `new` returns an array of shape `(0,num_columns)` while `current` returns an array of shape `(0,1)` and prints a `UserWarning: loadtxt: Empty input file`.


#### Copyright and Licensing

Please list the copyright holder for the work you are submitting (this will be you or your assignee, such as a university or company): University of Stuttgart

By submitting this pull request, the copyright holder is agreeing to license the submitted work under the following licenses:
- Code: BSD 3-clause (https://opensource.org/licenses/BSD-3-Clause)