@ahartikainen 2022-01-11
What are the runtimes when you have 100k columns and 100 draws?