@ahartikainen 2022-01-11
dtype should be forced to `float`

Is that timing table correct?