@qres 2022-01-11
The last two files are 30000 rows or 60000 rows x 236 columns.

Changing to `dtype=np.float64` did not change a lot:
| |49KB | 2.2MB | 66MB |
|--|-- | -- | -- |
|current with float64 |5.79 ms | 243 ms | 2.53 s |

> Is that timing table correct?

No. I swapped the row names. They are correct now.

> 100k columns and 100 draws

Good point. Seems like pandas does not like many columns even though the file size is still 87M. It took about 1min which is much more than the current numpy version wich takes 4.1s .

Therefore, I think it is best to keep everything as it is.