@ahartikainen 2022-01-11
Yeah, it would be really cool to find out why pandas is so slow for multiple columns.

I used initially pandas in arviz (for cmdstan) but in the end, for huge tables (huge # of columns) pandas was just too slow compared to basic python + numpy solution.

I think currently CmdStanPy has ok (ish) solution for reading large csv files. For small the time difference is not too big.

In future, a dask solution would be something that might be best for us.