@ahartikainen 2022-01-11
Oh, btw for a future reference, that forced float is due to stuck chains (e.g. stuck at init which has been set to integer value) -> merging int and float might cause problems. (of course I don't remember how it behaves when we put that data into a numpy array directly)